# contracts/openapi.yaml
openapi: 3.0.3
info:
  title: AetherQuant API Contract
  version: 1.0.0
  description: High-performance API Specification for AetherQuant AI Quant Research & Trading Platform
servers:
  - url: /api/v1
    description: Integrated Gateway & Worker API
paths:
...
  /instruments/search:
    get:
      summary: Search stocks & indices across CN/US
      parameters:
        - name: q
          in: query
          required: true
          schema: { type: string }
        - name: market
          in: query
          schema: { type: string, enum: [ALL, CN, US] }
      responses:
        '200':
          description: List of matched instruments
...
  /quotes/{symbol}:
    get:
      summary: Get realtime snapshot quote for instrument
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Realtime quote
  /bars/{symbol}:
    get:
      summary: Historical OHLCV bars (qfq/hfq)
      parameters:
        - name: symbol
          in: path
          required: true
          schema: { type: string }
        - name: period
          in: query
          schema: { type: string, default: "1M" }
        - name: adjust
          in: query
          schema: { type: string, default: "qfq" }
      responses:
        '200':
          description: OHLCV bars array
...